Got rid of the magic with the rel attribute.
The addNestedItem and removeNestedItem functions were looking at the
href and rel attributes to work out what elements they should be acting
on. Apart from being an abuse of those attributes, it was really pushing
functionality down into the HTML that shouldn't have been there.

Now the addNestedItem adn removeNestedItem functions are generators.
They take the elements to be acted on, and return functions that can be
used as click handlers. Much neater.